Update deploy script for air quality
Rearranged how BlobEntry objects are created - now added to a list as 
soon as available, instead of maintaining a lot of locals for filenames 
and content. Hence it's easy to loop over any number of markdown files 
and convert them all, in particular the air quality documentation as 
required.

Also a small bugfix: HTML pages now specify UTF-8 encoding, because 
there are UTF-8 characters in README_data but browsers were defaulting 
to plain ASCII.